bug: issue with refactor not re-mounting ShiftsViewVM
After refactor, the `ShiftsViewVM` will not re-render to alllow for guards to function correctly.  Instead of expected behavior, the view will not re-run the save function after pressing the save button, which holds state as it was when the button was pressed the first time.  Checking out old code to find the issue here.